지원한 서포터 목록 조회 API 구현 (#335)
* 오류나는 Dockerfile 수정 (#295)

* refactor: 필요없는 러너 서포터 필드 삭제

* refactor: 필요없는 ChattingCount 필드 삭제

* feat: RunnerTechnicalTag 관련 클래스 추가

* refactor: SupporterTechnicalTag 빌더 생성자 접근제한자 private 으로 변경

* feat: SupporterRunnerPost 엔티티 생성

* refactor: dev CI 스크립트 변경

* refactor: dev CD 스크립트 변경

* refactor: deploy CI 스크립트 변경

* refactor: deploy CD 스크립트 변경

* Auth ArgumentResolver 익명 null 반환 변경 (#271)

* refactor: Auth ArgumentResolver 익명 null 반환 변경

* refactor: AuthSupporterPrincipal 내부 속성 변경

* refactor: docker container timezone 변경

* feat: Dockerfile 타임존 설정

* feat: jar 타임존 설정

* fix: Dockerfile 오류 해결

---------


* feat: 지원한 서포터 목록 조회 API 구현

* fix: conflict 해결

* refactor: for 문 삭제

* test: path Parameter 추가

* fix: 충돌 해결

* refactor: introduction 객체 수정

* refactor: 코드 포맷팅

* style: 중복되는 테스트 삭제

* fix: 충돌 해결

* refactor: restdocs request Headers 추가

* refactor: 변수명 변경

* refactor: 변수명 변경

* refactor: 피드백 반영

* fix: introduction nullable=true 로 변경

* fix: introduction nullable=false로 변경

* refactor: 피드백 반영

* fix: 충돌 해결

* refactor: 피드백 반영

* refactor: 피드백 반영

* refactor: 피드백 반영

---------

+++ backend/baton/src/main/java/touch/baton/domain/runnerpost/controller/response/RunnerPostResponse.java
@@ -144,14 +144,14 @@ public record RunnerPostResponse() {
                                  LocalDateTime deadline,
                                  List<String> tags,
                                  int watchedCount,
-                                 int applicantCount,
+                                 long applicantCount,
                                  String reviewStatus
 
 
     ) {
 
         public static SimpleInMyPage from(final RunnerPost runnerPost,
-                                          final int applicantCount) {
+                                          final long applicantCount) {
             return new SimpleInMyPage(
                     runnerPost.getId(),
                     runnerPost.getTitle().getValue(),
